requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-RESTX==1.1.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.6
python-dotenv==1.0.0
Flask-Caching==2.0.2
Flask-Limiter==2.8.1
Flask-Testing==0.8.1
Flask-Swagger-UI==4.11.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
gunicorn==20.1.0